added chat and fixes
added signalR chat and a lot of fixes

+++ DonkeySellApi/Workers/CrudOnFavorites.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Data.Entity.Core;
 using System.Linq;
 using System.Threading.Tasks;
 using DonkeySellApi.Models;
@@ -20,10 +22,8 @@ namespace DonkeySellApi.Workers
         Task<List<Product>> GetUsersFavoriteProducts(string username);
     }
 
-    public class CrudOnFavorites: ICrudOnFavorites
+    public class CrudOnFavorites : ICrudOnFavorites, IDisposable
     {
-        // ufp = usersFavoriteProduct
-
         private DonkeySellContext context;
         private ICrudOnProducts crudOnProducts;
 
@@ -35,24 +35,27 @@ namespace DonkeySellApi.Workers
 
         public async Task<UsersFavoriteProducts> AddProductToFavorites(string username, int productId)
         {
-            if (! await ProductIsAllreadyAdded(username, productId))
-            {
-                var usf = new UsersFavoriteProducts() {ProductId = productId, Username = username};
-                context.UsersFavoriteProducts.Add(usf);
-                await context.SaveChangesAsync();
-                return usf;
-            }
+            if (!context.Users.Any(x => x.UserName == username))
+                throw new ObjectNotFoundException();
 
-            return null;
-        }
+            if (!context.Products.Any(x => x.Id == productId))
+                throw new ObjectNotFoundException();
 
-        private async Task<bool> ProductIsAllreadyAdded(string username, int productId)
-        {
-            return context.UsersFavoriteProducts.Any(x => x.ProductId == productId && x.Username == username);
+            if (context.UsersFavoriteProducts.Any(x => x.Username == username && x.ProductId == productId))
+                throw new FormatException();
+
+            var usf = new UsersFavoriteProducts() { ProductId = productId, Username = username };
+            context.UsersFavoriteProducts.Add(usf);
+            await context.SaveChangesAsync();
+            return usf;
+
         }
 
         public async Task<bool> DeleteProductFromFavorites(string username, int productId)
         {
+            if (!context.UsersFavoriteProducts.Any(x => x.Username == username && x.ProductId == productId))
+                throw new FormatException();
+
             var usf = context.UsersFavoriteProducts.Single(x => x.ProductId == productId && x.Username == username);
             context.UsersFavoriteProducts.Remove(usf);
             bool deleted = await context.SaveChangesAsync() > 0;
@@ -62,10 +65,13 @@ namespace DonkeySellApi.Workers
 
         public async Task<bool> DeleteProductFromAllUsers(int productId)
         {
+            if(!context.Products.Any(x => x.Id == productId))
+                throw new ObjectNotFoundException();
+
             var usfs = context.UsersFavoriteProducts.Where(x => x.ProductId == productId);
             usfs.ForEach(x =>
             {
-               context.UsersFavoriteProducts.Remove(x);
+                context.UsersFavoriteProducts.Remove(x);
             });
 
             await context.SaveChangesAsync();
@@ -75,6 +81,9 @@ namespace DonkeySellApi.Workers
 
         public async Task<List<Product>> GetUsersFavoriteProducts(string username)
         {
+            if(!context.Users.Any(x => x.UserName == username))
+                throw new ObjectNotFoundException();
+
             var usfs = context.UsersFavoriteProducts.Where(x => x.Username == username).ToList();
             var products = new List<Product>();
             usfs.ForEach(async x =>
@@ -84,6 +93,11 @@ namespace DonkeySellApi.Workers
             });
 
             return products;
+        }
+
+        public void Dispose()
+        {
+            context.Dispose();
         }
     }
 }
