+++ DonkeySellApi/Workers/CrudOnMessages.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data.Entity.Core;
 using System.Linq;
 using System.Threading.Tasks;
 using DonkeySellApi.Models;
@@ -19,12 +20,14 @@ namespace DonkeySellApi.Workers
         Task DeleteAllMessagesForProduct(int id);
 
         Task<List<Message>> GetUnreadMessages(string username);
+
         Task<Message> MessageWasRead(int id);
     }
 
-    public class CrudOnMessages: ICrudOnMessages
+    public class CrudOnMessages : ICrudOnMessages, IDisposable
     {
         private DonkeySellContext context;
+
         public CrudOnMessages()
         {
             context = new DonkeySellContext();
@@ -32,20 +35,30 @@ namespace DonkeySellApi.Workers
 
         public async Task<List<Message>> GetMessages(int productId)
         {
+            if (!context.Products.Any(x => x.Id == productId))
+                throw new ObjectNotFoundException();
+
             return context.Messages.Where(x => x.ProductId == productId).ToList();
         }
 
         public async Task<Message> AddOrUpdate(Message message)
         {
-            if (message.Id != 0 || context.Messages.Any(x => x.Id == message.Id))
-            {
-                var poco = context.Messages.Single(x => x.Id == message.Id);
-                poco.Value = message.Value;
-                poco.DateCreated = DateTime.Now;
-                await context.SaveChangesAsync();
+            if (message.Id == 0)
+                return await CreateMessage(message);
 
-                return poco;
-            }
+            if (!context.Messages.Any(x => x.Id == message.Id))
+                throw new ObjectNotFoundException();
+
+            return await UpdateMessage(message);
+        }
+
+        private async Task<Message> CreateMessage(Message message)
+        {
+            if (!message.IsValid())
+                throw new FormatException();
+
+            if(!context.Users.Any(x => x.UserName == message.UserName))
+                throw new ObjectNotFoundException();
 
             message.DateCreated = DateTime.Now;
             message.UserId = context.Users.Single(x => x.UserName == message.UserName).UserId;
@@ -55,8 +68,21 @@ namespace DonkeySellApi.Workers
             return message;
         }
 
+        private async Task<Message> UpdateMessage(Message message)
+        {
+            var poco = context.Messages.Single(x => x.Id == message.Id);
+            poco.Value = message.Value;
+            poco.DateCreated = DateTime.Now;
+            await context.SaveChangesAsync();
+
+            return poco;
+        }
+
         public async Task<int> DeleteMessage(int id)
         {
+            if(!context.Messages.Any(x => x.Id == id))
+                throw new ObjectNotFoundException();
+
             var poco = context.Messages.Single(x => x.Id == id);
             context.Messages.Remove(poco);
             await context.SaveChangesAsync();
@@ -66,6 +92,9 @@ namespace DonkeySellApi.Workers
 
         public async Task DeleteAllMessagesForProduct(int id)
         {
+            if(!context.Products.Any(x => x.Id == id))
+                throw new ObjectNotFoundException();
+
             var pocos = context.Messages.Where(x => x.ProductId == id);
             pocos.ForEach(async x =>
             {
@@ -75,17 +104,28 @@ namespace DonkeySellApi.Workers
 
         public async Task<List<Message>> GetUnreadMessages(string username)
         {
+            if (!context.Users.Any(x => x.UserName == username))
+                throw new ObjectNotFoundException();
+
             var pocos = context.Messages.Where(x => x.UserName == username && x.MessageWasRead == false);
             return pocos.ToList();
         }
 
         public async Task<Message> MessageWasRead(int id)
         {
+            if(!context.Messages.Any(x => x.Id == id))
+                throw new ObjectNotFoundException();
+
             var message = context.Messages.Single(x => x.Id == id);
             message.MessageWasRead = true;
             await context.SaveChangesAsync();
 
             return message;
+        }
+
+        public void Dispose()
+        {
+            context.Dispose();
         }
     }
 }
