+++ DonkeySellApi/Workers/CrudOnProducts.cs
@@ -1,9 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Data.Entity.Core;
 using System.Linq;
 using System.Threading.Tasks;
 using DonkeySellApi.Models;
 using DonkeySellApi.Models.DatabaseModels;
+using Microsoft.Data.OData.Query.SemanticAst;
+using Ninject.Infrastructure.Language;
 using WebGrease.Css.Extensions;
 
 namespace DonkeySellApi.Workers
@@ -19,7 +23,7 @@ namespace DonkeySellApi.Workers
         Task<List<Product>> GetProductsByQuery(string query, int take, int skip, string orderedBy);
     }
 
-    public class CrudOnProducts : ICrudOnProducts
+    public class CrudOnProducts : ICrudOnProducts, IDisposable
     {
         private DonkeySellContext context;
         private ICrudOnMessages crudOnMessages;
@@ -36,29 +40,30 @@ namespace DonkeySellApi.Workers
 
         public async Task<Product> GetProduct(int id)
         {
+            if (!context.Products.Any(x => x.Id == id))
+                throw new ObjectNotFoundException();
+
             return context.Products.Single(x => x.Id == id);
         }
 
         public async Task<Product> AddOrUpdate(Product product)
         {
-            if (product.Id != 0 || context.Products.Any(x => x.Id == product.Id))
-            {
-                var poco = context.Products.Single(x => x.Id == product.Id);
-                await ClearProductOfImages(product.Id);
+            if (!product.IsValid())
+                throw new FormatException();
 
-                poco.City = product.City;
-                poco.Category = product.Category;
-                poco.Title = product.Title;
-                poco.Price = product.Price;
-                poco.TradesAccepted = poco.TradesAccepted;
-                poco.Rental = poco.Rental;
-                poco.Free = poco.Free;
-                poco.Description = product.Description;
-                poco.Images = product.Images;
-                await context.SaveChangesAsync();
+            if (product.Id == 0)
+                return await CreateProduct(product);
 
-                return poco;
-            }
+            if (!context.Products.Any(x => x.Id == product.Id))
+                throw new ObjectNotFoundException();
+
+            return await UpdateProduct(product);
+        }
+
+        private async Task<Product> CreateProduct(Product product)
+        {
+            if(!context.Users.Any(x => x.UserName == product.UserName))
+                throw new ObjectNotFoundException();
 
             product.DatePublished = DateTime.Now;
             product.UserId = context.Users.Single(x => x.UserName == product.UserName).UserId;
@@ -66,6 +71,25 @@ namespace DonkeySellApi.Workers
             await context.SaveChangesAsync();
 
             return product;
+        }
+
+        private async Task<Product> UpdateProduct(Product product)
+        {
+            var poco = context.Products.Single(x => x.Id == product.Id);
+            await ClearProductOfImages(product.Id);
+
+            poco.City = product.City;
+            poco.Category = product.Category;
+            poco.Title = product.Title;
+            poco.Price = product.Price;
+            poco.TradesAccepted = poco.TradesAccepted;
+            poco.Rental = poco.Rental;
+            poco.Free = poco.Free;
+            poco.Description = product.Description;
+            poco.Images = product.Images;
+            await context.SaveChangesAsync();
+
+            return poco;
         }
 
         private async Task ClearProductOfImages(int id)
@@ -79,17 +103,32 @@ namespace DonkeySellApi.Workers
 
         public async Task<int> DeleteProduct(int id)
         {
+            if (!context.Products.Any(x => x.Id == id))
+                throw new ObjectNotFoundException();
+
             var poco = context.Products.Single(x => x.Id == id);
             await crudOnMessages.DeleteAllMessagesForProduct(id);
             await DeleteProductFromFavorites(id);
+            await DeleteImagesOfProduct(poco.Images.ToList());
             context.Products.Remove(poco);
             await context.SaveChangesAsync();
 
             return poco.Id;
         }
 
+        private async Task DeleteImagesOfProduct(List<Image> images)
+        {
+            images.ForEach(x =>
+            {
+                context.Images.Remove(x);
+            });
+        }
+
         public async Task DeleteAllProductsOfUser(string username)
         {
+            if (!context.Users.Any(x => x.UserName == username))
+                throw new ObjectNotFoundException();
+
             var pocos = context.Products.Where(x => x.UserName == username).ToList();
             pocos.ForEach(async p =>
             {
@@ -98,11 +137,14 @@ namespace DonkeySellApi.Workers
                 context.Products.Remove(p);
             });
 
-            await  context.SaveChangesAsync();
+            await context.SaveChangesAsync();
         }
 
         private async Task DeleteProductFromFavorites(int id)
         {
+            if (!context.Products.Any(x => x.Id == id))
+                throw new ObjectNotFoundException();
+
             var favorites = context.UsersFavoriteProducts.Where(x => x.ProductId == id).ToList();
             favorites.ForEach(p =>
             {
@@ -112,6 +154,9 @@ namespace DonkeySellApi.Workers
 
         public async Task<List<Product>> GetProductsOfUser(string username)
         {
+            if (!context.Users.Any(x => x.UserName == username))
+                throw new ObjectNotFoundException();
+
             var poco = context.Products.Where(x => x.UserName == username);
 
             return poco.ToList();
@@ -121,7 +166,30 @@ namespace DonkeySellApi.Workers
         {
             var products = query != null ? context.Products.SqlQuery(query).ToList() : context.Products.ToList();
 
-            return products.OrderByDescending(x => x.GetType().GetProperty(orderedBy).GetValue(x, null)).Skip(skip).Take(take).ToList();
+            var simpleProducts =
+                products.Select(
+                    x =>
+                        new Product()
+                        {
+                            Id = x.Id,
+                            Title = x.Title,
+                            Price = x.Price,
+                            UserName = x.UserName,
+                            Category = x.Category,
+                            DatePublished = x.DatePublished,
+                            City = x.City,
+                            Images = x.Images.Any() ? new Collection<Image>() {x.Images.ToList()[0]} : null
+                        });
+            return
+                simpleProducts.OrderByDescending(x => x.GetType().GetProperty(orderedBy).GetValue(x, null))
+                    .Skip(skip)
+                    .Take(take)
+                    .ToList();
+        }
+
+        public void Dispose()
+        {
+            context.Dispose();
         }
     }
 }
